// datasync-clientes/src/types/Costumer.ts
export type Customer = {
    id: number;
    corporate: boolean;
    name: string;
    trade_name?: string;
    person_identification_number: string;
    state_inscription?: string;
    commercial_address?: string;
    commercial_address_number?: string;
    business_district?: string;
    commercial_zip_code?: string;
    billingPhone?: string;
    email: string;
    email_nfe?: string;
    customer_origin?: string;
    final_customer: boolean;
    billing_id?: string;
    square_id?: number;
    activity_id?: number;
    business_city?: string;
    seller_id?: number;
    city_id?: number;
    country_id?: number;
    document_type?: string;
    registered: boolean;
    filial: number;
    rca: string;
    novo: number;
    atualizado: number;
    data_nascimento: string;
    created_at?: Date;
    updated_at?: Date;
};